# vulkan: Implement "fast divide" (mul+shift) for unary ops like copy