Integer division is relatively expensive on GPUs. Division by a constant can be done with a mul+shift of some precomputed values (See https://gmplib.org/~tege/divcnst-pldi94.pdf figure 4.1).

This change uses the fast divide for the coordinate calculations in generic_unary_head.comp, primarily intended to speed up noncontiguous copy shaders. Copies are still relatively expensive in some models, and I can see a 1-2% speedup in some cases (with coopmat) with this change. I also added a couple test-backend-ops perf tests that benefit (measured on RTX 4070):

```
before:
  CPY(type_src=f32,type_dst=f32,ne=[8192,512,2,1],permute=[0,2,1,3]):                   6669 runs -   160.82 us/run -    65536 kB/run -  389.02 GB/s
  CPY(type_src=f32,type_dst=f32,ne=[3072,512,2,1],permute=[0,2,1,3]):                  23222 runs -    43.37 us/run -    24576 kB/run -  540.66 GB/s
after:
  CPY(type_src=f32,type_dst=f32,ne=[8192,512,2,1],permute=[0,2,1,3]):                   7182 runs -   148.61 us/run -    65536 kB/run -  420.98 GB/s
  CPY(type_src=f32,type_dst=f32,ne=[3072,512,2,1],permute=[0,2,1,3]):                  42346 runs -    24.18 us/run -    24576 kB/run -  969.74 GB/s
cuda:
  CPY(type_src=f32,type_dst=f32,ne=[8192,512,2,1],permute=[0,2,1,3]):                   5130 runs -   195.87 us/run -    65536 kB/run -  319.41 GB/s
  CPY(type_src=f32,type_dst=f32,ne=[3072,512,2,1],permute=[0,2,1,3]):                  15026 runs -    70.60 us/run -    24576 kB/run -  332.12 GB/s
```